requirements/base.txt
aiofiles==23.1.0
aioredis==2.0.1
hiredis==2.2.2
alabaster==0.7.12
#amqp==5.1.1
appnope==0.1.3
asgiref==3.6.0
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8==1.4.4
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.12.0
billiard==3.6.4.0
biopy-isatab==0.1.1
biopython==1.73
bleach==3.1.0
boto3==1.26.96
botocore==1.29.96
#Brotli==1.0.9
bs4==0.0.1
celery==5.2.7
certifi==2019.6.16
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==3.0.4
colorama==0.4.1
colored==1.4.3
configparser==3.7.4
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
-e git+https://github.com/IQSS/dataverse-client-python.git@5f95a11e2f32bcfabe28e6ad43b2f7edac60d983#egg=dataverse
decorator==4.4.0
deepdiff==4.0.7
defusedxml==0.6.0
Deprecated==1.2.13
dict-toolbox==2.2.0
dict.sorted==1.0.0
Django==4.1.7
django-allauth==0.52.0
django-allow-cidr==0.6.0
django-appconf==1.0.5
django-compression-middleware==0.4.2
django-compressor==4.3.1
django-cors-headers==3.14.0
django-extensions==3.2.1
django-gollum==1.0.1
django-html-validator==0.5.1
django-redis-sessions==0.6.2
django-tools==0.54.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
djangorestframework==3.14.0
docopt==0.6.2
docutils==0.12

entrypoints==0.3
enum-compat==0.0.2
et-xmlfile==1.0.1
eventlet==0.22.1
Genshi==0.7.3
gevent==22.10.2
greenlet==2.0.1
gunicorn==19.9.0
#hiredis==1.0.0
html5lib==1.1
hurry.filesize==0.9
hyperlink==21.0.0
idna==2.6
imagesize==0.7.1
importlib-metadata==4.13.0
incremental==22.10.0
#ipykernel==5.1.2
#ipython==7.8.0
#ipython-genutils==0.2.0
#ipywidgets==7.5.1
#isatools==0.14.2
iso8601==0.1.12
isodate==0.6.0
isort==4.2.15
jdcal==1.3
jedi==0.15.1
Jinja2==3.1.2
jmespath==0.9.0
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
jsonpickle==1.2
jsonref==0.2
jsonschema==2.6.0
#jupyter==1.0.0
#jupyter-client==5.3.3
#jupyter-console==6.0.0
#jupyter-core==4.5.0
kombu==5.2.4
lazy-object-proxy==1.3.1
lxml==4.9.1
MarkupSafe==2.0.1
mccabe==0.6.1
meld3==2.0.0
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
msgpack==1.0.3
mzml2isa==0.5.1
#nbconvert==5.6.0
#nbformat==4.4.0
nest-asyncio==1.5.5
netaddr==0.8.0
networkx==2.0
#notebook==6.0.1
numexpr==2.8.1
numpy==1.24.1
oauthlib==2.0.6
openpyxl==3.1.2
ordered-set==3.1.1
packaging==21.3
pandas==1.5.3
pandocfilters==1.4.2
parso==0.5.1
pbr==5.8.0
pexpect==4.4.0
pickleshare==0.7.5
ply==3.9
pop-config==12.0.1
pop-loop==1.0.5
progressbar2==3.34.3
#prometheus-client==0.7.1
prompt-toolkit==2.0.10
pronto==0.10.1
psycopg2-binary==2.9.5
ptyprocess==0.5.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==1.7.2
pymongo==3.12.3
pyOpenSSL==23.0.0
pyparsing==2.1.10
PyPDF2==1.26.0
pyrsistent==0.15.4
pytest==3.0.4
python-dateutil==2.8.2
python-utils==2.3.0
python3-openid==3.0.10
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
#qtconsole==4.5.5
rauth==0.7.3
rcssmin==1.1.1
rdflib==6.2.0
redis==4.2.0
#rend==6.4.2
requests==2.28.0
requests-oauthlib==0.7.0
rjsmin==1.2.1
rsa==3.4.2
s3transfer==0.6.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
snowballstemmer==1.2.1
soupsieve==1.9.2
sphinxcontrib-websupport==1.0.1
sqlparse==0.4.2
supervisor==4.2.4
tables==3.7.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.2
tornado==6.0.3
#traitlets==4.3.3
Twisted==22.10.0
txaio==23.1.1
typed-ast==1.5.4
typing==3.6.2
urllib3==1.26.15
validators==0.20.0
vine==5.0.0
wcwidth==0.1.7
webencodings==0.5.1
#widgetsnbextension==3.5.1
wrapt==1.10.10
xlrd==1.2.0
xmljson==0.1.9
XlsxWriter==3.0.3
zipp==0.6.0
zope.interface==5.5.2
Pillow==9.3.0
tzdata==2022.7
